Minor tweaks
Removed some unnecessary "using" statements.

The right panel's drop shadow now adapts to theme changes for better visibility when the dark mode is active.

Fixed a bug that caused the UI to remain in the small state when maximized albeit enough space being available to switch to the full view.