Notifications and rewritten responsive design
Notifications also respect the corresponding setting.

Decreased visual glitches caused by the responsive design (especially when an image is visible).

Rewrote the code responsible for the responsive design.

The scan button itself is now showing whether a scan is in progress to treat all display sizes fairly.

Users with small screens now also have the ability to edit a scan. When they are done, they can click a button and return to the left pane.

Spread out text a little more on the settings page to make it less busy.

Added an automatic version indicator to the settings page.