File conversion, improved error handling and much more
Unsupported image formats can now be used if a base image format is available. The related setting is respected as well and can be changed during runtime.

"Ctrl+S" and "Ctrl+C" trigger share or copy if a result is visible.

Many many errors especially related to file operations (and permissions) are now caught.

A new method "refreshLeftPanel()" is now taking care of the left panel in most situations. It includes some (perhaps unnecessary) redundancies for possible UI anomalies. "scanner_selected()" and "no_scanner_selected()" have been replaced.

Changing the source mode doesn't reset the selected file format anymore (if the format is still available in the new mode).

The app now has a new logo. The taskbar icons support the dark and light theme. The badge icons are still missing and need to be designed.

Website link in settings now correctly triggers the default browser.

The save location option looks a bit more organized and can be reset to its default value. If the "Scans" folder doesn't exist it is created.

Added a new flow state for the upcoming inking feature.

The app now tracks which version it had the last time it was launched to detect app updates and perhaps show a changelog in the future.

Copying an image to the clipboard now triggers a rich toast notification.

Fixed the new scan button's text color to be visible regardless of the theme.