Error handling and cancel feature
All scanner errors should now be correctly caught and processed. An error message appears when something goes wrong and a notification is sent if the app is in the background while the error occurs.

A new cancel button appears during a scan and disappears when the scan ends for any reason.

Further tweaks to the spacing of the settings page and drop shadows in the app.

Hid the draw button for now.

Added a new method LocalizedString() that allows to get localized strings and make the code easier to read.