Rework of the secondary CommandBar
The secondary CommandBar is now the only regular way to exit the cropping mode. It is more context-aware and allows the user to "discard", "save" and "save as copy".

The aspect ratio flyout now includes the option to flip the current aspect ratio.

The aspect ratio flyout now features rounded corners for a more consistent appearance.

Updated the icon for the (currently hidden) draw button to better represent its purpose.

Added some translatable strings.