Minor improvements
Switching from "Flatbed"/"Feeder" to "Automatic configuration" is now faster on the surface.

Fix for the text in the right panel not reappearing if a nwely selected scanner fails.

The current file name is now selected when opening the rename ContentDialog for easy replacement.

Removed debugging line previously pushed by accident.

Selected a different icon for the recents button to better depict its purpose.

Fixed a bug related to the new ContentDialog, which caused a rename to happen twice when pressing the enter key.