Preparation for first release
Improved keyboard shortcuts to be more reliable.

Loading the scanFolder is more reliable.

Added missing Reveal effects to the save button.

Fixed a bug preventing the app from setting the scanFolder on first launch.

Fixed a bug causing the text in the right panel to reappear prematurely if the last scanner is lost.