"Open with" and some improvements
There is now a new "Open with" button, which displays the app picker and allows the user to quickly open the current file with the app of their choice.

Drawing is now possible with mouse & pen.

Users now see a notification about the save location when they first launch the app.

Errors related to the save location that can occur while scanning are neatly displayed with a TeachingTip.

Fixed an error which caused unsupported source modes to become available when an error occurred.

Fixed an error causing the app to claim that the scanner returned an error when in reality the target folder just didn't exist.