Experimental PDF conversion added
The solution now includes a win32 component called "ImageToPDF", which can be used to convert a scan to a PDF file using PDFsharp.
The app now requires fullTrust permissions to run the win32 component.

If a compatible source format is available, the conversion algorithm now offers PDF as a new file type.