PDF fixes and other enhancements
Fixed some issues with the hand-off process from win32 to UWP. Unfortunately, this includes a delay that can be up to 3 seconds long.

Fixed the numbering algorithm for converted PDFs.

Added a debug shortcut (Ctrl + D) that allows the manual selection of a result file (image only for now).

Migrated many format strings to a new Enum called "SupportedFormat" for safer coding.

Decreased the margin of the CommandBars.

Updated the changelog.

Added the PDFsharp license link.

The aspect ratio button now displays the currently selected item.