Reworked the search indicator
The indicator above the scanner list and its corresponding setting and strings have been removed. There is now a permanent indicator inside the scanner ComboBox, which delivers more information on what exactly is happening and makes the overall UI look a lot cleaner.

Made sure that no format is selected anymore when the last scanner in the list is lost.