Foundation for v2.0
The MainPage has been completely rewritten and features a new responsive design for small, normal and wide displays.

New scan options have been added for duplex scan and for scanning multiple pages with the feeder configuration.

An indicator icon is shown next to source modes that support preview. Previewing is available as an option in the new scan button's flyout.

A new panel called "Organize" has been added. It collects all scans in a session. A new session can be started manually with the new scan button, or by selecting a different file format. The different pages can be reorganized using drag&drop. This will be especially useful for changing the order of the pages of a PDF, for other file types this doesn't have any practical use.

The scanner list now uses the new class "RecognizedScanner", which allows the app to collect all info on a scanner as soon as it is detected. This should increase performance and reliability when switching between scanners.

Added some slick animations to multiple parts of the interface.

Added a debug menu that can be accessed by selecting the settings button while pressing CTRL.

The app doesn't allow native PDF scanning anymore and will instead use conversion in any case for consistent results.

In its current state, the app lacks the necessary UI for editing scans. Additionally, no conversion is supported.

The localization is mostly broken.