Major features added and bugs fixed
Refined the Organize pane and added a TextBlock at the top, which will be used to display information on the current session.

Added the new link indicators to the MainPage.

The toolbar is now enabled once a scan is present.

The Copy, OpenWith and Share buttons in the toolbar have been implemented.

Fixed some visual glitches when the app is running inside a very tight window.

Fixed some threading stuff in ScanResult and also made extensive preparations for the reimplementation of PDF conversion.

Removed ScanInCorrectMode() as it's no longer needed.

InitializeTempFolder() now deletes all files in the temp directory AND its sub directories.