PDF functionality mostly done + many fixes/changes
Rebuilt ImageToPDF to be able to construct the new multi-page files.

Implemented a new scope TeachingTip that appears when the user wants copy/open/share an element in a generated document (i.e. a PDF file).

Refined layout appearance of the Organize panel.

Organizing a scan is now called "Managing a scan".

Made sure that the titlebar text is correct even for PDF files and when a scan is dismissed.

Switched to a ProgressRing on the PreviewPage for better consistency.

Fixed some threading and other stuff in ScanResult mostly related to PDFs.

Fixed some bugs.