GridView for PDF, simulated auto preview, debug PDF, fixes
The management pane switches to a GridView for PDF files. Both ListView and GridView now block reordering in select mode.

If auto preview is not supported and only one other mode (flatbed / feeder) is available and this other mode supports preview, the app will allow the user to preview the auto configuration. This will simply make use of the only available preview mode.

The format ComboBox in the debug menu is now used to choose the desired target format of the selected image files.
This also supports PDF conversion.
If necessary, the CheckBox for starting fresh will be forcibly selected.

Fixed the shadow width in the management pane if the window is small.

Fixed an issue that caused rotating multiple images simultaneously to fail with certain selections.

Removed XPS and OXPS from format selection.