Crop scan, append time to file name
Scans can now be cropped and the changes can be saved to the selected file or as a copy.

Added a CheckBox in settings, which causes a timestamp ("_HHMMSS") to be appended to the file names.

Prevented an exception occuring in ImageToPDF to terminate close the entire app environment.

Added a toolbar for inking.