Draw on scan, improvements, fixes
Drawing on scans now works as expected.

Made sure that the appropriate code runs on the UI thread in SettingsPage.xaml.cs.
Added an animation to the settings header when entering the page.
Errors in settings now use a TeachingTip instead of a ContentDialog.
Removed the setting for automatically switching to drawing mode when a pen is detected.

Fixed a bug that caused the source mode to always be set to "Automatic" when a scan is completed.

Added a success animation to saving edits as a copy.

Newly discovered scanners are now added to the bottom of the list (just above the search indicator).