Cancel scan, translations & much more
The cancel button that appears when a scan is in progress now works as expected.

Created localization resources from scratch for the entire application and added German and English resources.

Fixed ProgressRing getting stuck after a page navigation.

Improved the way the error message is displayed on the preview page.

Added a tutorial for page management which appears the first time 2 or more pages have been scanned for natural discovery.

Removed the setting for automatic scanner selection - no longer needed with the more resilient way of discovering scanners in v2.0.

Added credits to the bottom of the settings page.

Implemented keyboard shortcut for copying a scan.

The app heading "Scanner" is now hidden when there isn't much space.

Improved many English strings.

Added a subtle launch animation.

Fixed a lot of stuff.