Fixes and improvements
Fixed a bug that caused a crash when generating an existing PDF again.

Improved the high-contrast theme and added high-contrast variants of all visual assets.

Made it more obvious whether the app is currently active or not.

Event logs are now exported as .txt for easier upload to GitHub.

Improved display of save location in the page list.