Added scan overlay
The new overlay displays the current page number, the total amount of pages, and offers zoom controls.

Refined the appearance of the custom toolbar buttons.

Improved RTL support.

The log file will now be finalized prior to a crash to prevent data loss.

Prevented the rename textbox from growing in width and consequently breaking the TeachingTip.