Numerous improvements
Started removing old components.

The page list is now locked during a scan.

The scan button now displays an icon to indicate whether pages will be added to an existing result or a new result will be created.

Added a print button to the page list.

The SettingsView now has a maximum width to avoid stretching the content too much.

Fixed a bug causing the page list to be missing when the app is launched in a wide state.