Improved opening in another app
All pages will now be dismissed if the user selects one of the suggested apps or launches the system dialog. A warning is displayed, which can be dismissed each time or set to not appear again.

Fixed titlebar button color for when the system theme is overridden by the app setting.

Added file conversion when scanning.

Removed the old DisplayedFolder property in ScanResultElement.