Improved scan and merge
The scan and merge dialog is now much more accessible than before.
If the current scanner is lost, the dialog will be closed immediately.
The elements in the merge preview are now easier to distinguish.

Scan and merge into PDF is now also available in auto mode, if the scanner supports feeder scanning.