Unit tests & virtual brightness/contrast defaults
Improved the brightness and contrast configuration of DiscoveredScanners, as well as the property discovery. The default brightness and contrast are now replaced with a very close virtual default value to work around a bug in the Windows Scan Runtime.

Added some unit tests for this new logic.

Improved localization around the upcoming scan region selection.